Reverted Sequelize files to Javascript Sequelize CLI does not support Typescript yet
