+++ src/database/models/business_units.js
@@ -1,0 +1,24 @@
+"use strict";
+const {
+    Model
+} = require("sequelize");
+module.exports = (sequelize, DataTypes) => {
+    class business_units extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+        static associate(models) {
+            // define association here
+        }
+    }
+    business_units.init({
+        id: DataTypes.INTEGER,
+        name: DataTypes.STRING
+    }, {
+        sequelize,
+        modelName: "business_units",
+    });
+    return business_units;
+};
