+++ src/database/models/payments_periods.js
@@ -1,0 +1,24 @@
+"use strict";
+const {
+    Model
+} = require("sequelize");
+module.exports = (sequelize, DataTypes) => {
+    class payments_periods extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+        static associate(models) {
+            // define association here
+        }
+    }
+    payments_periods.init({
+        id: DataTypes.INTEGER,
+        name: DataTypes.STRING
+    }, {
+        sequelize,
+        modelName: "payments_periods",
+    });
+    return payments_periods;
+};
