+++ src/database/models/privileges.js
@@ -1,0 +1,24 @@
+"use strict";
+const {
+    Model
+} = require("sequelize");
+module.exports = (sequelize, DataTypes) => {
+    class privileges extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+        static associate(models) {
+            // define association here
+        }
+    }
+    privileges.init({
+        id: DataTypes.INTEGER,
+        description: DataTypes.STRING
+    }, {
+        sequelize,
+        modelName: "privileges",
+    });
+    return privileges;
+};
