--- src/database/seeders/20220524165630-demo-business_units.ts
@@ -1,16 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up (queryInterface, Sequelize) {
-        await queryInterface.bulkInsert("business_units", [{
-            name: "Michelada",
-        },
-        {
-            name: "Arkus Nexus",
-        }]);
-    },
-
-    async down (queryInterface, Sequelize) {
-        await queryInterface.bulkDelete("business_units", null, {});
-    }
-};
