--- src/database/seeders/20220525160030-demo-payments_periods.ts
@@ -1,17 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up (queryInterface, Sequelize) {
-        await queryInterface.bulkInsert("payments_periods", [{
-            name: "Quincenal",
-        },
-        {
-            name: "Mensual",
-        }]);
-    },
-
-    async down (queryInterface, Sequelize) {
-        await queryInterface.bulkDelete("payments_periods", null, {});
-    }
-
-};
