+++ src/database/seeders/20220525162203-demo-catalogs.js
@@ -1,0 +1,15 @@
+"use strict";
+
+module.exports = {
+    async up (queryInterface, Sequelize) {
+        await queryInterface.bulkInsert("catalogs", [{
+            name: "Edición",
+            description: "Puede editar",
+            privileges: {privileges:[123,456]} , 
+        }],{}, {privileges:{type:new Sequelize.JSON()}});
+    },
+
+    async down (queryInterface, Sequelize) {
+        await queryInterface.bulkDelete("catalogs", null, {});
+    }
+};
