+++ src/database/seeders/20220525170849-demo-privileges.js
@@ -1,0 +1,16 @@
+"use strict";
+
+module.exports = {
+    async up (queryInterface, Sequelize) {
+        await queryInterface.bulkInsert("privileges", [{
+            description: "Crear administradores",
+        },
+        {
+            description: "Crear unidades de negocio",
+        }]);
+    },
+
+    async down (queryInterface, Sequelize) {
+        await queryInterface.bulkDelete("privileges", null, {});
+    }
+};
