--- src/database/seeders/20220525170849-demo-privileges.ts
@@ -1,16 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up (queryInterface, Sequelize) {
-        await queryInterface.bulkInsert("privileges", [{
-            description: "Crear administradores",
-        },
-        {
-            description: "Crear unidades de negocio",
-        }]);
-    },
-
-    async down (queryInterface, Sequelize) {
-        await queryInterface.bulkDelete("privileges", null, {});
-    }
-};
