--- src/database/seeders/20220526170144-demo-payrolls.ts
@@ -1,16 +1,0 @@
-"use strict";
-
-module.exports = {
-    async up (queryInterface, Sequelize) {
-        await queryInterface.bulkInsert("payrolls", [{
-            payment_date: "2022-06-01",
-            payment_period_id: 1,
-            business_unit_id: 2,
-            total_amount: 12354.23
-        },]);
-    },
-
-    async down (queryInterface, Sequelize) {
-        await queryInterface.bulkDelete("payrolls", null, {});
-    }
-};
